Core: Move SI source files to their own directory
