+++ Source/Core/Core/Core.vcxproj
@@ -140,14 +140,14 @@
     <ClCompile Include="HW\MemoryInterface.cpp" />
     <ClCompile Include="HW\MMIO.cpp" />
     <ClCompile Include="HW\ProcessorInterface.cpp" />
-    <ClCompile Include="HW\SI.cpp" />
-    <ClCompile Include="HW\SI_Device.cpp" />
-    <ClCompile Include="HW\SI_DeviceDanceMat.cpp" />
-    <ClCompile Include="HW\SI_DeviceGBA.cpp" />
-    <ClCompile Include="HW\SI_DeviceGCAdapter.cpp" />
-    <ClCompile Include="HW\SI_DeviceGCController.cpp" />
-    <ClCompile Include="HW\SI_DeviceGCSteeringWheel.cpp" />
-    <ClCompile Include="HW\SI_DeviceKeyboard.cpp" />
+    <ClCompile Include="HW\SI\SI.cpp" />
+    <ClCompile Include="HW\SI\SI_Device.cpp" />
+    <ClCompile Include="HW\SI\SI_DeviceDanceMat.cpp" />
+    <ClCompile Include="HW\SI\SI_DeviceGBA.cpp" />
+    <ClCompile Include="HW\SI\SI_DeviceGCAdapter.cpp" />
+    <ClCompile Include="HW\SI\SI_DeviceGCController.cpp" />
+    <ClCompile Include="HW\SI\SI_DeviceGCSteeringWheel.cpp" />
+    <ClCompile Include="HW\SI\SI_DeviceKeyboard.cpp" />
     <ClCompile Include="HW\Sram.cpp" />
     <ClCompile Include="HW\StreamADPCM.cpp" />
     <ClCompile Include="HW\SystemTimers.cpp" />
@@ -370,14 +370,14 @@
     <ClInclude Include="HW\MMIO.h" />
     <ClInclude Include="HW\MMIOHandlers.h" />
     <ClInclude Include="HW\ProcessorInterface.h" />
-    <ClInclude Include="HW\SI.h" />
-    <ClInclude Include="HW\SI_Device.h" />
-    <ClInclude Include="HW\SI_DeviceDanceMat.h" />
-    <ClInclude Include="HW\SI_DeviceGBA.h" />
-    <ClInclude Include="HW\SI_DeviceGCAdapter.h" />
-    <ClInclude Include="HW\SI_DeviceGCController.h" />
-    <ClInclude Include="HW\SI_DeviceGCSteeringWheel.h" />
-    <ClInclude Include="HW\SI_DeviceKeyboard.h" />
+    <ClInclude Include="HW\SI\SI.h" />
+    <ClInclude Include="HW\SI\SI_Device.h" />
+    <ClInclude Include="HW\SI\SI_DeviceDanceMat.h" />
+    <ClInclude Include="HW\SI\SI_DeviceGBA.h" />
+    <ClInclude Include="HW\SI\SI_DeviceGCAdapter.h" />
+    <ClInclude Include="HW\SI\SI_DeviceGCController.h" />
+    <ClInclude Include="HW\SI\SI_DeviceGCSteeringWheel.h" />
+    <ClInclude Include="HW\SI\SI_DeviceKeyboard.h" />
     <ClInclude Include="HW\Sram.h" />
     <ClInclude Include="HW\StreamADPCM.h" />
     <ClInclude Include="HW\SystemTimers.h" />
