+++ Source/Core/Core/Core.vcxproj.filters
@@ -445,28 +445,28 @@
     <ClCompile Include="HW\ProcessorInterface.cpp">
       <Filter>HW %28Flipper/Hollywood%29\PI - Processor Interface</Filter>
     </ClCompile>
-    <ClCompile Include="HW\SI.cpp">
+    <ClCompile Include="HW\SI\SI.cpp">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClCompile>
-    <ClCompile Include="HW\SI_Device.cpp">
+    <ClCompile Include="HW\SI\SI_Device.cpp">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClCompile>
-    <ClCompile Include="HW\SI_DeviceDanceMat.cpp">
+    <ClCompile Include="HW\SI\SI_DeviceDanceMat.cpp">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClCompile>
-    <ClCompile Include="HW\SI_DeviceGBA.cpp">
+    <ClCompile Include="HW\SI\SI_DeviceGBA.cpp">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClCompile>
-    <ClCompile Include="HW\SI_DeviceGCAdapter.cpp">
+    <ClCompile Include="HW\SI\SI_DeviceGCAdapter.cpp">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClCompile>
-    <ClCompile Include="HW\SI_DeviceGCController.cpp">
+    <ClCompile Include="HW\SI\SI_DeviceGCController.cpp">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClCompile>
-    <ClCompile Include="HW\SI_DeviceGCSteeringWheel.cpp">
+    <ClCompile Include="HW\SI\SI_DeviceGCSteeringWheel.cpp">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClCompile>
-    <ClCompile Include="HW\SI_DeviceKeyboard.cpp">
+    <ClCompile Include="HW\SI\SI_DeviceKeyboard.cpp">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClCompile>
     <ClCompile Include="HW\VideoInterface.cpp">
@@ -1042,28 +1042,28 @@
     <ClInclude Include="HW\ProcessorInterface.h">
       <Filter>HW %28Flipper/Hollywood%29\PI - Processor Interface</Filter>
     </ClInclude>
-    <ClInclude Include="HW\SI.h">
+    <ClInclude Include="HW\SI\SI.h">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClInclude>
-    <ClInclude Include="HW\SI_Device.h">
+    <ClInclude Include="HW\SI\SI_Device.h">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClInclude>
-    <ClInclude Include="HW\SI_DeviceDanceMat.h">
+    <ClInclude Include="HW\SI\SI_DeviceDanceMat.h">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClInclude>
-    <ClInclude Include="HW\SI_DeviceGBA.h">
+    <ClInclude Include="HW\SI\SI_DeviceGBA.h">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClInclude>
-    <ClInclude Include="HW\SI_DeviceGCAdapter.h">
+    <ClInclude Include="HW\SI\SI_DeviceGCAdapter.h">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClInclude>
-    <ClInclude Include="HW\SI_DeviceGCController.h">
+    <ClInclude Include="HW\SI\SI_DeviceGCController.h">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClInclude>
-    <ClInclude Include="HW\SI_DeviceGCSteeringWheel.h">
+    <ClInclude Include="HW\SI\SI_DeviceGCSteeringWheel.h">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClInclude>
-    <ClInclude Include="HW\SI_DeviceKeyboard.h">
+    <ClInclude Include="HW\SI\SI_DeviceKeyboard.h">
       <Filter>HW %28Flipper/Hollywood%29\SI - Serial Interface</Filter>
     </ClInclude>
     <ClInclude Include="HW\VideoInterface.h">
