+++ Source/Core/Core/HW/SI/SI.cpp
@@ -2,6 +2,8 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
+#include "Core/HW/SI/SI.h"
+
 #include <algorithm>
 #include <array>
 #include <cstring>
@@ -13,8 +15,7 @@
 #include "Core/CoreTiming.h"
 #include "Core/HW/MMIO.h"
 #include "Core/HW/ProcessorInterface.h"
-#include "Core/HW/SI.h"
-#include "Core/HW/SI_DeviceGBA.h"
+#include "Core/HW/SI/SI_DeviceGBA.h"
 #include "Core/HW/SystemTimers.h"
 #include "Core/Movie.h"
 #include "Core/NetPlayProto.h"
