+++ Source/Core/Core/HW/SI/SI_DeviceGCSteeringWheel.cpp
@@ -2,7 +2,8 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
-#include "Core/HW/SI_DeviceGCSteeringWheel.h"
+#include "Core/HW/SI/SI_DeviceGCSteeringWheel.h"
+
 #include "Common/CommonTypes.h"
 #include "Common/Logging/Log.h"
 #include "Core/HW/GCPad.h"
