+++ Source/Core/Core/HW/SI/SI_DeviceGCSteeringWheel.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "Core/HW/SI_DeviceGCController.h"
+#include "Core/HW/SI/SI_DeviceGCController.h"
 
 class CSIDevice_GCSteeringWheel : public CSIDevice_GCController
 {
