+++ Source/Core/Core/HW/SI/SI_DeviceKeyboard.cpp
@@ -2,7 +2,8 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
-#include "Core/HW/SI_DeviceKeyboard.h"
+#include "Core/HW/SI/SI_DeviceKeyboard.h"
+
 #include "Common/ChunkFile.h"
 #include "Common/CommonTypes.h"
 #include "Common/Logging/Log.h"
