+++ Source/Core/InputCommon/GCAdapter_Android.cpp
@@ -13,7 +13,7 @@
 #include "Core/ConfigManager.h"
 #include "Core/Core.h"
 #include "Core/CoreTiming.h"
-#include "Core/HW/SI.h"
+#include "Core/HW/SI/SI.h"
 #include "Core/HW/SystemTimers.h"
 
 #include "InputCommon/GCAdapter.h"
